파일 분리 및 AuditionManager 타입 수정
