+++ CodeStarterCamp_Week5/Level.swift
@@ -1,0 +1,14 @@
+//
+//  Level.swift
+//  CodeStarterCamp_Week5
+//
+//  Created by Hyejeong Jeong on 2022/11/15.
+//
+
+import Foundation
+
+enum Level {
+    case A
+    case B
+    case C
+}
