+++ CodeStarterCamp_Week5/Person.swift
@@ -1,0 +1,18 @@
+//
+//  Person.swift
+//  CodeStarterCamp_Week5
+//
+//  Created by Hyejeong Jeong on 2022/11/15.
+//
+
+import Foundation
+
+class Person {
+    let name: String
+    var height: Int
+    
+    init(name: String, height: Int) {
+        self.name = name
+        self.height = height
+    }
+}
